# openapi.yaml
openapi: 3.0.3
info:
  title: Last Price
  description: |
    Server for conducting A/B testing experiments on pricing strategies using Last Price integration.
    
    This server provides endpoints for:
    - Assigning users to experiment variants
    - Tracking pricing page views and conversions
    - Processing Paid.ai webhook events
    - Analyzing experiment results
    
    **Architecture**: The system uses deterministic SHA-256 hashing to ensure consistent user-to-variant mapping,
    emits signals to Paid.ai's Signals API for event tracking, and provides real-time analytics.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /debug/assignments:
    get:
      summary: Get all experiment assignments
      description: |
        Returns all current experiment assignments for debugging purposes.
        Only available in development mode.
      tags:
        - Debug
      responses:
        '200':
          description: Assignments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 10
                  assignments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Assignment'
        '403':
          description: Endpoint only available in development mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============ Last Price API Paths ============
  /customers:
    get:
      tags:
        - Customers
      summary: List all customers
      description: Retrieves all customers from Last Price
      operationId: listLastPriceCustomers
      responses:
        '200':
          description: Retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LastPriceCustomer'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Customers
      summary: Create a customer
      description: Creates a new customer in Last Price
      operationId: createLastPriceCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceCustomerInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceCustomer'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /invoices/{invoiceId}:
    get:
      tags:
        - Invoices
      summary: Retrieve an invoice
      description: Retrieves an invoice by ID from Last Price
      operationId: getLastPriceInvoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceInvoice'

    put:
      tags:
        - Invoices
      summary: Update an invoice
      description: Updates an invoice in Last Price
      operationId: updateLastPriceInvoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceInvoiceInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceInvoice'

    delete:
      tags:
        - Invoices
      summary: Delete an invoice
      description: Deletes an invoice from Last Price
      operationId: deleteLastPriceInvoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
  /invoices/{invoiceId}/send:
    post:
      tags:
        - Invoices
      summary: Send an invoice
      description: Sends an invoice to the customer
      operationId: sendLastPriceInvoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Invoice sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceInvoice'
...
  /disputes:
    get:
      tags:
        - Disputes
      summary: List all disputes
      description: Retrieves all disputes from Last Price
      operationId: listLastPriceDisputes
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, under_review, won, lost]
      responses:
        '200':
          description: Retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LastPriceDispute'

    post:
      tags:
        - Disputes
      summary: Create a dispute
      description: Creates a new dispute in Last Price
      operationId: createLastPriceDispute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceDisputeInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceDispute'
  /disputes/{disputeId}:
    get:
      tags:
        - Disputes
      summary: Retrieve a dispute
      description: Retrieves a dispute by ID from Last Price
      operationId: getLastPriceDispute
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceDispute'

    put:
      tags:
        - Disputes
      summary: Update a dispute
      description: Updates a dispute in Last Price
      operationId: updateLastPriceDispute
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceDisputeInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceDispute'
...
components:
  schemas:
...
    Assignment:
      type: object
      properties:
        userId:
          type: string
          example: user_123
        experimentId:
          type: string
          example: pricing_test_001
        variant:
          type: string
          enum: [control, experiment]
          example: control
...
    LastPriceAddress:
      type: object
      properties:
        line1:
          type: string
          example: "123 Business Ave"
        line2:
          type: string
          example: "Suite 100"
        city:
          type: string
          example: "San Francisco"
        state:
          type: string
          example: "CA"
        zipCode:
          type: string
          example: "94105"
        country:
          type: string
          example: "USA"
    LastPriceCustomer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
          example: "Acme Corporation"
        phone:
          type: string
          example: "+1-555-0123"
        website:
          type: string
          example: "https://acme.com"
        employeeCount:
          type: integer
          example: 100
        annualRevenue:
          type: number
          example: 1000000
        taxExemptStatus:
          type: string
          enum: [none, exempt, reverse]
          default: none
        creationSource:
          type: string
          enum: [manual, api, crm, other]
        externalId:
          type: string
          example: "acme_123"
        creationState:
          type: string
          example: "active"
        billingAddress:
          $ref: '#/components/schemas/LastPriceAddress'
    LastPriceCustomerInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        phone:
          type: string
        website:
          type: string
        employeeCount:
          type: integer
        annualRevenue:
          type: number
        taxExemptStatus:
          type: string
          enum: [none, exempt, reverse]
        creationSource:
          type: string
          enum: [manual, api, crm, other]
        externalId:
          type: string
        billingAddress:
          $ref: '#/components/schemas/LastPriceAddress'
...
    LastPriceInvoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        invoiceNumber:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        issueDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        currency:
          type: string
          example: USD
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/LastPriceInvoiceLineItem'
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LastPriceInvoiceInput:
      type: object
      required:
        - customerId
        - lineItems
      properties:
        customerId:
          type: string
          format: uuid
        issueDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/LastPriceInvoiceLineItem'
        notes:
          type: string
        currency:
          type: string
          default: USD
    LastPriceInvoiceLineItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        amount:
          type: number

    # ============ Payment Schemas ============
...
    LastPriceDispute:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        paymentId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [open, under_review, won, lost]
        reason:
          type: string
        evidence:
          type: string
        customerMessage:
          type: string
        resolution:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
    LastPriceDisputeInput:
      type: object
      required:
        - paymentId
        - reason
      properties:
        paymentId:
          type: string
          format: uuid
        reason:
          type: string
        evidence:
          type: string
        customerMessage:
          type: string

    # ============ Credit Schemas ============
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request parameters
        message:
          type: string
          example: userId must be a non-empty string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for Last Price API authentication
    webhookSignature:
      type: apiKey
      in: header
      name: x-paid-signature
      description: HMAC signature for webhook verification